+++ src/main/java/step3/Controller.java
@@ -1,6 +1,6 @@
 package step3;
 
-import step3.service.Racing;
+import step3.domain.Racing;
 import step3.view.InputView;
 import step3.view.ResultView;
