+++ src/main/java/step3/domain/Cars.java
@@ -1,6 +1,6 @@
 package step3.domain;
 
-import step3.service.strategy.MoveStrategy;
+import step3.domain.strategy.MoveStrategy;
 
 import java.util.List;
 import java.util.Objects;
