+++ src/main/java/step3/domain/Randomize.java
@@ -1,4 +1,4 @@
-package step3.service;
+package step3.domain;
 
 import java.util.Random;
