+++ src/test/java/step3/RacingTest.java
@@ -2,7 +2,7 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.service.Racing;
+import step3.domain.Racing;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
